JobberCRMFlask/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
psycopg2-binary==2.9.7
redis==4.6.0
gunicorn==21.2.0
PyJWT==2.8.0
requests==2.31.0
marshmallow==3.20.1